requirements.txt
alembic==1.12.1
aniso8601==9.0.1
APScheduler==3.10.4
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-mysql-connector==1.1.0
flask-restplus==0.13.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pluggy==1.3.0
protobuf==4.21.12
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytest==7.4.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
waitress==2.1.2
Werkzeug==3.0.0
wheel==0.41.2
